Because perl needs to know when it successfully loaded a module:
It does that by checking if the module returned a true value.

To do this, we just make sure the last statement in a module is true. "1"
is a true value that became a convention because that's exactly what it
refers to, but any true value will do, such as a string, or a different
number. ("true" values in Perl is anything that isn't undef, empty string,
or the number zero)

Some modules find it funny to end with "OMG WHAT THE FUCK" as the last
line. It works because that string is obviously true (because it's not
empty). Some modules find is useful (no idea why, honestly) to end it with
the name of the module:

    package MyModule;

    ...

    # last line, instead of 1;
    "MyModule";

There is a post about it here: http://www.perlmonks.org/?node_id=781320,
and it lists a few funny examples:

- 42;
- 8;
- "true";
- "false";
- "this is the end of the module";
- "I was there 05/07/04";

All of these are true values (even the string "false", because it's not
an empty string, just a string that says "false"), so all of these work.

The following website has a list of modules with weird return codes,
including some really funny ones:

http://returnvalues.useperl.at/