This can simply be subroutine-scoped:
When used outside the subroutine, the scope is the file itself, which
makes it accessible to every subroutine and piece of code in the file.
That's definitely not what we want.

I'm assuming it was like that before a clean up, and it was simply missed
once the original code was cleaned.