Use "defined or" syntax from v5.10:
Perl version 5.10 added an operator called "defined-or", which exists for
situations like this. It looks like "//=". This is similar to "||=" which
means "true-or".

The problem with "true-or" is that it will override defined values which
are false. If you gave it zero, even though zero is false (as a boolean),
you don't want to override it. You actually want to check if it's defined,
not if it's true, right?

So:

    if ( ! defined $value ) {
        $value = 'default';
    }

    # same as:
    $value = 'default' unless defined $value;

    # same as:
    $value //= 'default';